+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingTabs/proposalVotingTabs.test.tsx
@@ -1,12 +1,16 @@
 import { render, screen } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { type RefObject } from 'react';
+import { ProposalVotingStatus } from '../../proposalUtils';
 import { ProposalVotingTab } from '../proposalVotingDefinitions';
 import { type IProposalVotingTabsProps, ProposalVotingTabs } from './proposalVotingTabs';
 
 describe('<ProposalVotingTabs /> component', () => {
     const createTestComponent = (props?: Partial<IProposalVotingTabsProps>) => {
-        const completeProps: IProposalVotingTabsProps = { ...props };
+        const completeProps: IProposalVotingTabsProps = {
+            status: ProposalVotingStatus.ACTIVE,
+            ...props,
+        };
 
         return <ProposalVotingTabs {...completeProps} />;
     };
@@ -47,4 +51,16 @@ describe('<ProposalVotingTabs /> component', () => {
         await userEvent.click(screen.getByRole('tab', { name: 'Details' }));
         expect(screen.getByRole('tab', { name: 'Breakdown' })).toBeInTheDocument();
     });
+
+    it.each([{ status: ProposalVotingStatus.PENDING }, { status: ProposalVotingStatus.UNREACHED }])(
+        'disables the breakdown and votes tabs when voting status is $status',
+        ({ status }) => {
+            render(createTestComponent({ status }));
+            expect(screen.getByRole('tab', { name: 'Breakdown' }).getAttribute('disabled')).toEqual('');
+            expect(screen.getByRole('tab', { name: 'Votes' }).getAttribute('disabled')).toEqual('');
+
+            const detailsTab = screen.getByRole('tab', { name: 'Details' });
+            expect(detailsTab.getAttribute('disabled')).toBeNull();
+        },
+    );
 });